requirements.txt
langchain_openai
langchain_core
langchain_community
python-dotenv
streamlit
langserve
fastapi
uvicorn
sse_starlette
pytesseract
bs4
pypdf
pandas
numpy
faiss-cpu